Feat(SSR): change RestController to Controller and add products.html for view all Products
add init DataSet in Application.java
