+++ src/main/java/gift/controller/ProductController.java
@@ -1,20 +1,15 @@
 package gift.controller;
 
+
 import gift.domain.Product;
-import gift.dto.CreateProductRequest;
-import gift.dto.CreateProductResponse;
-import gift.dto.UpdateProductRequest;
-import gift.dto.UpdateProductResponse;
 import gift.service.ProductService;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
-
-@RestController
+@Controller
 @RequestMapping("/api/products")
 public class ProductController {
 
@@ -24,39 +19,10 @@ public class ProductController {
         this.service = service;
     }
 
-    @PostMapping
-    public HttpEntity<CreateProductResponse> createProduct(@RequestBody CreateProductRequest request) {
-        CreateProductResponse createProductResponse = service.save(request);
-
-        return new ResponseEntity<>(createProductResponse, HttpStatus.CREATED);
-    }
-
-    @GetMapping("/{id}")
-    public HttpEntity<Product> findProductById(@PathVariable Long id) {
-        Product findProduct = service.findById(id);
-
-        return new ResponseEntity<>(findProduct, HttpStatus.OK);
-    }
-
     @GetMapping
-    public HttpEntity<List<Product>> findProducts() {
+    public String items(Model model) {
         List<Product> products = service.findAll();
-
-        return new ResponseEntity<>(products, HttpStatus.OK);
-    }
-
-    @PutMapping("/{id}")
-    public HttpEntity<UpdateProductResponse> updateProduct(@PathVariable Long id, @RequestBody UpdateProductRequest request) {
-        UpdateProductResponse updateProductResponse = service.update(id, request);
-
-        return new ResponseEntity<>(updateProductResponse, HttpStatus.OK);
-
-    }
-
-    @DeleteMapping("/{id}")
-    public HttpEntity<Void> deleteProduct(@PathVariable Long id) {
-        service.delete(id);
-
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        model.addAttribute("products", products);
+        return "/products";
     }
 }
