+++ src/main/resources/templates/products.html
@@ -1,0 +1,52 @@
+<!DOCTYPE HTML>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="utf-8">
+  <link href="../css/bootstrap.min.css"
+        th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
+</head>
+<body>
+
+<div class = "container" style="max-width: 600px">
+  <div class="py-4 text-center">
+    <h2>상품 관리</h2>
+  </div>
+
+  <div class="row">
+    <div class="col">
+      <button class="btn btn-dark float-end" th:onclick="|location.href='@{/api/products/add}'|" type="button"> 상품 등록</button>
+    </div>
+  </div>
+
+  <hr class="my-4">
+  <div>
+    <table class="table">
+      <thead>
+      <tr>
+        <th>ID</th>
+        <th>상품명</th>
+        <th>가격</th>
+        <th>ImageURL</th>
+        <th>수정</th>
+        <th>삭제</th>
+
+      </tr>
+      </thead>
+      <tbody>
+      <tr th:each="product: ${products}">
+        <td th:text="${product.id}">10000</td>
+        <td th:text="${product.name}">10000</td>
+        <td th:text="${product.price}">10000</td>
+        <td th:text="${product.imageUrl}">example.com</td>
+        <td><a th:href="@{/api/products/{productId}/edit(prductId=${product.id})}">수정</a></td>
+        <td><a th:href="@{/api/products/{productId}/delete(prductId=${product.id})}">삭제</a></td>
+
+      </tr>
+      </tbody>
+    </table>
+
+  </div>
+</div>
+
+</body>
+</html>
